refactor: rename computeMissingVersions file
