+++ dailyBuild/fetchVersionsHistory.ts
